requirements.txt
flask
flask-cors
pandas
numpy
python-dotenv
alpaca-trade-api
v20